requirements.txt
pillow==3.0.0
uniseg